requirements.txt
flask==1.0.2
psycopg2==2.9.1
python-decouple==3.5